requirements.txt
aldjemy==2.6
attrs==21.2.0
databases==0.5.3
Django==2.2.10
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
greenlet==1.1.2
iniconfig==1.1.1
packaging==21.2
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pydantic==1.8.2
pyhumps==3.0.2
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.3
SQLAlchemy==1.4.27
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
